requirements.txt
langchain 
langchain-community 
langchain-text-splitters 
langchain-google-genai 
pypdf 
faiss-cpu 
python-dotenv